apps/analysis-engine/requirements.txt
pyspark==3.5.4
psycopg2-binary==2.9.10
python-dotenv==1.1.0
